+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v0.0.6 (04/25/2020)
 
-- Auto-commits are now only made when changes files don't have any associated errors. This behavior can be changed with the new `GitDoc: Ignore Errors` setting.
+- Auto-commits are now only made when changes files don't have any associated errors. This behavior can be changed with the new `GitDoc: Commit Validation Level` setting.
 
 ## v0.0.5 (04/23/2020)